Bump packages (#23)
* feat: Add ^ to package.json dependencies

* * *
The vite config now is clear,
With react we banish all fear.
No more will we see,
A bug wild and free,
Our code is now perfectly sheer.
* * *

* feat: Upgrade dependencies

* * *
The tailwindcss styles now gleam,
A fresh new design is the theme.
With react so bright,
Our future's in sight,
A beautiful, functional dream.
* * *

* feat: Upgrade vite to v7

Viteをv7にアップグレードしました。

* * *
The new vite starts to run,
With speed and grace for everyone.
A faster build, a quicker load,
A better developer road.
* * *